refactor: split normalized stamp and my-page repositories
