+++ backend/app/repository.py
@@ -2,9 +2,9 @@
 
 from __future__ import annotations
 
-from .repository_normalized import get_or_create_user
 from .repository_support import format_datetime, to_seoul_date, utcnow_naive
 from .repositories.legacy_stamp_repository import get_stamps, toggle_stamp
+from .repositories.user_data_repository import get_or_create_user
 
 __all__ = [
     "format_datetime",
